requirements.txt
streamlit
python-dotenv
langchain-core
langchain-community
huggingface-hub